Rename OpenAIRealtimeAssistant component file
